+++ packages/fighting-design/text/index.ts
@@ -8,7 +8,7 @@ export type TextInstance = InstanceType<typeof Text>
 
 export type { TextPropsType } from './src/text'
 
-export type { TextType } from './src/interface'
+export type { TextType, TextDecorationType } from './src/interface'
 
 declare module 'vue' {
   export interface GlobalComponents {