+++ start/src/App.vue
@@ -1,7 +1,12 @@
 <script lang="ts" setup></script>
 
 <template>
-  <f-button type=""></f-button>
+  <f-text indent="20px" line-height="30px" center background="">
+    背景颜色
+  </f-text>
+  <f-text background="#e372ff">背景颜色</f-text>
+  <f-text background="#e49b33">背景颜色</f-text>
+  <p>233</p>
 </template>
 
 <style lang="scss" scoped></style>